Intel-Systems: Cleanup Coverity identified issues and address range issues

+++ Intel-Systems/common/ieprom1.c
@@ -136,7 +136,7 @@ uint8 EPROM1_get_mbyte(uint16 addr)
 {
     uint8 val;
 
-    if ((addr >= EPROM1_unit->u3) && ((uint32) addr <= (EPROM1_unit->u3 + EPROM1_unit->capac))) {
+    if ((addr >= EPROM1_unit->u3) && ((uint16) addr <= (EPROM1_unit->u3 + EPROM1_unit->capac))) {
         SET_XACK(1);                /* good memory address */
         val = *((uint8 *)EPROM1_unit->filebuf + (addr - EPROM1_unit->u3));
         val &= 0xFF;
